pyproject.toml
[project]
name = "prism-cli"
version = "0.1.0"
description = "PRISM — Project Reasoning & Intelligent Skill Memory"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-frontmatter>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
fase1 = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.24",
    "tiktoken",
]
fase2 = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "watchdog>=3.0",
    "httpx",
]
fase4 = [
    "anthropic",
    "gitpython>=3.0",
]

[project.scripts]
prism = "prism.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["prism/**"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
